Make `ReadoutType` non-generic
