+++ Mlem/App/Enums/Interaction/CommentBarConfiguration+Types.swift
@@ -61,7 +61,7 @@ extension CommentBarConfiguration {
             }
         }
         
-        func associatedReadouts(context: any InteractableProviding) -> Set<Configuration.ReadoutType> {
+        func associatedReadouts(context: any InteractableProviding) -> Set<ReadoutType> {
             switch self {
             case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
             case .downvote: context.votes.value?.myVote ?? .none == .downvote ? [.downvote, .score] : [.downvote]
@@ -109,7 +109,7 @@ extension CommentBarConfiguration {
             }
         }
         
-        func associatedReadouts(context: any InteractableProviding) -> Set<Configuration.ReadoutType> {
+        func associatedReadouts(context: any InteractableProviding) -> Set<ReadoutType> {
             switch self {
             case .score: [.upvote, .downvote, .score]
             case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
@@ -118,33 +118,4 @@ extension CommentBarConfiguration {
             }
         }
     }
-    
-    enum ReadoutType: String, ReadoutTypeProviding {
-        case created
-        case score
-        case upvote
-        case downvote
-        case comment
-        case saved
-        
-        var appearance: MockReadoutAppearance {
-            switch self {
-            case .created: .init(icon: .general.time, label: "18h")
-            case .score: .init(icon: .lemmy.votes, label: "7")
-            case .upvote: .init(icon: .lemmy.upvoted, label: "9")
-            case .downvote: .init(icon: .lemmy.downvoted, label: "2")
-            case .comment: .init(icon: .lemmy.replies, label: "1")
-            case .saved: .init(icon: .lemmy.saved, label: "")
-            }
-        }
-        
-        func compatibleWith(otherReadouts: Set<Self>) -> Bool {
-            switch self {
-            case .score: otherReadouts.isDisjoint(with: [.upvote, .downvote])
-            case .upvote, .downvote: !otherReadouts.contains(.score)
-            default: true
-            }
-        }
-    }
-
 }
