+++ Mlem/App/Enums/Interaction/PostBarConfiguration+Types.swift
@@ -69,16 +69,7 @@ extension PostBarConfiguration {
             }
         }
         
-        func associatedReadouts(context: any InteractableProviding) -> Set<PostBarConfiguration.ReadoutType> {
-            switch self {
-            case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
-            case .downvote: context.votes.value?.myVote ?? .none == .downvote ? [.downvote, .score] : [.downvote]
-            case .save: [.saved]
-            case .reply, .share, .selectText, .postDetails, .hide, .block, .report, .crossPost, .lock, .pin, .resolve, .remove, .ban: []
-            }
-        }
-        
-        func associatedReadouts(context: Post) -> Set<PostBarConfiguration.ReadoutType> {
+        func associatedReadouts(context: any InteractableProviding) -> Set<ReadoutType> {
             switch self {
             case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
             case .downvote: context.votes.value?.myVote ?? .none == .downvote ? [.downvote, .score] : [.downvote]
@@ -128,16 +119,7 @@ extension PostBarConfiguration {
             }
         }
         
-        func associatedReadouts(context: any InteractableProviding) -> Set<PostBarConfiguration.ReadoutType> {
-            switch self {
-            case .score: [.upvote, .downvote, .score]
-            case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
-            case .downvote: context.votes.value?.myVote ?? .none == .downvote ? [.downvote, .score] : [.downvote]
-            case .reply: []
-            }
-        }
-        
-        func associatedReadouts(context: Post) -> Set<PostBarConfiguration.ReadoutType> {
+        func associatedReadouts(context: any InteractableProviding) -> Set<ReadoutType> {
             switch self {
             case .score: [.upvote, .downvote, .score]
             case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
@@ -146,33 +128,4 @@ extension PostBarConfiguration {
             }
         }
     }
-    
-    enum ReadoutType: String, ReadoutTypeProviding {
-        case created
-        case score
-        case upvote
-        case downvote
-        case comment
-        case saved
-        
-        var appearance: MockReadoutAppearance {
-            switch self {
-            case .created: .init(icon: .general.time, label: "18h")
-            case .score: .init(icon: .lemmy.votes, label: "7")
-            case .upvote: .init(icon: .lemmy.upvoted, label: "9")
-            case .downvote: .init(icon: .lemmy.downvoted, label: "2")
-            case .comment: .init(icon: .lemmy.replies, label: "1")
-            case .saved: .init(icon: .lemmy.saved, label: "")
-            }
-        }
-        
-        func compatibleWith(otherReadouts: Set<Self>) -> Bool {
-            switch self {
-            case .score: otherReadouts.isDisjoint(with: [.upvote, .downvote])
-            case .upvote, .downvote: !otherReadouts.contains(.score)
-            default: true
-            }
-        }
-    }
-
 }
