+++ Mlem/App/Utility/Extensions/Content Models/Comment/Comment+Actions.swift
@@ -13,7 +13,7 @@ import SwiftUI
 extension Comment {
     // MARK: - Readouts
     
-    func readout(type: CommentBarConfiguration.ReadoutType, showColor: Bool) -> Readout? {
+    func readout(type: ReadoutType, showColor: Bool) -> Readout? {
         switch type {
         case .created: createdReadout
         // swiftlint:disable:next void_function_in_ternary
@@ -25,18 +25,6 @@ extension Comment {
         }
     }
 
-    func readout(type: ReplyBarConfiguration.ReadoutType, showColor: Bool) -> Readout? {
-        switch type {
-        case .created: createdReadout
-        // swiftlint:disable:next void_function_in_ternary
-        case .score: downvotesEnabled ? scoreReadout(showColor: showColor) : upvoteReadout(showColor: showColor)
-        case .upvote: upvoteReadout(showColor: showColor)
-        case .downvote: downvotesEnabled ? downvoteReadout(showColor: showColor) : nil
-        case .comment: commentReadout
-        case .saved: savedReadout(showColor: showColor)
-        }
-    }
-    
     // MARK: - Counters
     
     func counter(
